+++ GHEtool/Borefield.py
@@ -565,8 +565,7 @@ class Borefield(BaseClass):
             Equivalent borehole thermal resistance [mK/W]
         """
         return self.borehole.get_Rb(self.H, self.D, self.r_b, self.ground_data.k_s, self.depth,
-                                    temperature=min(self.Tf_min,
-                                                    self.results.min_temperature if self.results.min_temperature is not None else 10e6))
+                                    temperature=self.results.min_temperature if self.results.min_temperature is not None else self.Tf_min)
 
     @Rb.setter
     def Rb(self, Rb: float) -> None:
@@ -1937,7 +1936,8 @@ class Borefield(BaseClass):
         self.custom_gfunction = CustomGFunction(time_array, borehole_length_array, options)
         self.custom_gfunction.create_custom_dataset(self.borefield, self.ground_data.alpha)
 
-    def Re(self, **kwargs) -> float:
+    @property
+    def Re(self) -> float:
         """
         Reynolds number.
 
@@ -1946,7 +1946,8 @@ class Borefield(BaseClass):
         float
             Reynolds number
         """
-        return self.borehole.Re(**kwargs)
+        return self.borehole.Re(
+            temperature=self.results.min_temperature if self.results.min_temperature is not None else self.Tf_min)
 
     def calculate_quadrant(self) -> int:
         """