+++ GHEtool/VariableClasses/FluidData/CommercialFluids/__init__.py
@@ -1,0 +1,2 @@
+from .ThermoxDTX import ThermoxDTX
+from .CoolflowNTP import CoolflowNTP