+++ GHEtool/VariableClasses/FluidData/TemperatureDependentFluidData.py
@@ -8,6 +8,8 @@ from scp.water import Water
 from scp.base_melinder import BaseMelinder
 
 from GHEtool.VariableClasses.FluidData._FluidData import _FluidData
+from GHEtool.VariableClasses.FluidData.CommercialFluids._CommercialFluids import _CommercialFluids
+from GHEtool.VariableClasses.FluidData.CommercialFluids import *
 from GHEtool.VariableClasses.FluidData.ConstantFluidData import ConstantFluidData
 from GHEtool.VariableClasses.BaseClass import BaseClass
 from typing import Union
@@ -38,18 +40,22 @@ class TemperatureDependentFluidData(_FluidData, BaseClass):
         self._name = name
         self._percentage = percentage
         self._mass_percentage = mass_percentage
-        self._fluid: BaseMelinder | None = None
+        self._fluid: BaseMelinder | _CommercialFluids | None = None
 
         super().__init__(0)
 
         self.set_fluid(name, percentage)
 
-        # create interp array
-        self._spacing = np.linspace(self.freezing_point, 100, TemperatureDependentFluidData.NB_OF_SAMPLES)
-        self._k_f_array = np.array([self._fluid.conductivity(i) for i in self._spacing])
-        self._mu_array = np.array([self._fluid.viscosity(i) for i in self._spacing])
-        self._rho_array = np.array([self._fluid.density(i) for i in self._spacing])
-        self._cp_array = np.array([self._fluid.specific_heat(i) for i in self._spacing])
+        if isinstance(self._fluid, _CommercialFluids):
+            self._spacing = self._fluid._temperatures[::-1]
+        else:
+            # create interp array
+            self._spacing = np.linspace(self.freezing_point, 100, TemperatureDependentFluidData.NB_OF_SAMPLES)
+
+        self._k_f_array = np.array([float(self._fluid.conductivity(i)) for i in self._spacing])
+        self._mu_array = np.array([float(self._fluid.viscosity(i)) for i in self._spacing])
+        self._rho_array = np.array([float(self._fluid.density(i)) for i in self._spacing])
+        self._cp_array = np.array([float(self._fluid.specific_heat(i)) for i in self._spacing])
 
     def _calc_density_antifreeze(self, percentage: float = 30) -> float:
         """
@@ -145,6 +151,12 @@ class TemperatureDependentFluidData(_FluidData, BaseClass):
         if name == 'MEA':
             fluid = EthylAlcohol(percentage / 100)
             return fluid, fluid.freeze_point(percentage / 100)
+        if name == 'Thermox DTX':
+            fluid = ThermoxDTX(percentage / 100)
+            return fluid, fluid.freeze_point(percentage / 100)
+        if name == 'Coolflow NTP':
+            fluid = CoolflowNTP(percentage / 100)
+            return fluid, fluid.freeze_point(percentage / 100)
         raise ValueError(f'The fluid {name} is not yet supported by GHEtool.')
 
     def set_fluid(self, name: str, percentage: float) -> None:
@@ -166,9 +178,9 @@ class TemperatureDependentFluidData(_FluidData, BaseClass):
         if name in ['Water', 'MPG', 'MEG', 'MMA', 'MEA'] and not self._mass_percentage:
             # these properties are saved as mass percentage so the percentage should be converted
             percentage = self._convert_to_mass_percentage(percentage)
-        # if name not in ['Water', 'MPG', 'MEG', 'MMA', 'MEA'] and self._mass_percentage:
-        #     # these properties are saved as volume percentage so the percentage should be converted
-        #     percentage = self._convert_to_vol_percentage(percentage)
+        if name not in ['Water', 'MPG', 'MEG', 'MMA', 'MEA'] and self._mass_percentage:
+            # these properties are saved as volume percentage so the percentage should be converted
+            percentage = self._convert_to_vol_percentage(percentage)
 
         self._fluid, self._freezing_point = self._get_fluid(name, percentage)